ADDING TO ADMIN DASHBOARD

+++ api/utils/seeds.js
@@ -1,51 +1,45 @@
 require('dotenv').config();
 const mongoose = require('mongoose');
-const bcrypt = require('bcrypt');
 const User = require('../models/User');
+const bcrypt = require('bcrypt');
 
 const saltRounds = 10;
 
 const users = [
-  { username: 'admin', password: 'admin123' },
-  { username: 'john', password: 'pass3' },
-  { username: 'jane', password: 'pass4' },
-  { username: 'josh', password: 'pass6' },
-  { username: 'jake', password: 'pass7' },
+  { firstname: 'John', lastname: 'Doe', email: 'john@example.com', username: 'john', password: 'password123' },
+  { firstname: 'Jane', lastname: 'Doe', email: 'jane@example.com', username: 'jane', password: 'password123' },
 ];
+
 
 async function seed() {
   try {
-    await mongoose.connect(process.env.MONGODB_URI, {
+    // Load the MongoDB URI from environment variables
+    const uri = process.env.MONGODB_URI;
+
+    if (!uri) {
+      throw new Error('MONGODB_URI environment variable is not set');
+    }
+
+    await mongoose.connect(uri, {
       useNewUrlParser: true,
       useUnifiedTopology: true,
     });
 
     console.log('Connected to MongoDB');
 
+
     // Clear existing collections
     await User.deleteMany({});
 
     // Hash passwords before inserting users
     const hashedUsers = await Promise.all(users.map(async user => ({
-      username: user.username,
+      ...user,
       password: await bcrypt.hash(user.password, saltRounds),
     })));
 
-    // Insert users and get user IDs
-    const insertedUsers = await User.insertMany(hashedUsers);
-    const userMap = insertedUsers.reduce((map, user) => {
-      map[user.username] = user._id;
-      return map;
-    }, {});
+    // Insert users
+    await User.insertMany(hashedUsers);
 
-    // Update the admin user to have isAdmin set to true
-    await User.updateOne(
-      { username: 'admin' },
-      { $set: { isAdmin: true } }
-    );
-
-    console.log('Users inserted:', insertedUsers);
-    console.log('Admin user updated to isAdmin: true');
     console.log('Database seeded successfully');
   } catch (err) {
     console.error('Error seeding database:', err);
